requirements.txt
jupyter==1.0.0
matplotlib==2.1.0
numpy==1.13.3
Pillow==4.3.0
tensorflow==1.3.0
tflearn==0.3.2
tqdm==4.19.4
